Add print mode
Add support for printShader.

When the entry point of the shader code is "printMain", we will switch
to the print mode, where the canvas element will be hidden, and will
will be replaced by a readonly textarea element.

In this mode, we will only render for one frame, and the outputBuffer
will be copied back to CPU and formatted to a string, then output to
the textarea.